requirements.txt
PyYAML==3.10
Flask==0.10.1
boto==2.25.0
